move logic to split lines into test assembly
